Fix horrible bug!
There was a terrible bug, I was in Java mode. I was using class variables when
I should have been using instance variables. Which meant that all objects
properties were being shared in some cases, like the choices dictionary.

So, this fixes that and also updates the README to show the correct data.